Put all sound things into one sound js
